+++ GameServer/quests/Quests/QuestScripts/albion/epic/Defenders50.cs
@@ -102,7 +102,7 @@ namespace DOL.GS.Quests.Albion
                     log.Warn("Could not find "+Lidmann.Name+", creating it ...");
 
 				Lidmann.GuildName = "";
-				Lidmann.Realm = (byte) eRealm.Albion;
+				Lidmann.Realm = eRealm.Albion;
 				Lidmann.CurrentRegionID = 1;
 				Lidmann.Size = 50;
 				Lidmann.Level = 50;
@@ -131,7 +131,7 @@ namespace DOL.GS.Quests.Albion
 				Uragaig.Model = 349;
 				Uragaig.Name = "Cailleach Uragaig";
 				Uragaig.GuildName = "";
-				Uragaig.Realm = (byte) eRealm.None;
+				Uragaig.Realm = eRealm.None;
 				Uragaig.CurrentRegionID = 1;
 				Uragaig.Size = 55;
 				Uragaig.Level = 70;
